requirements.txt
jupyter 
ipykernel
pandas
matplotlib

dotenv
pyyaml
regex

gurobipy
coptpy
pyscipopt

openai
google
google-genai
google-generativeai

sentence-transformers 
hdbscan
scikit-learn

datasets
omegaconf


openai>=1.0.0
requests>=2.32.2
pandas>=2.2.2
rich>=13.3.5
gurobipy>=12.0.2